// src/shared/types/nutrition.tsx
export type Nutrition = {
    calories: string | null
    carbohydrates: string | null
    cholesterol: string | null
    fat: string | null
    fiber: string | null
    protein: string | null
    saturatedFat: string | null
    sodium: string | null
    sugar: string | null
    transFat: string | null
    unsaturatedFat: string | null
}